Updated Hand Detection to use model output
updated hand_detection to work with model output rather than the xmls we are using the train the model.

For the first test case, we can now generate a pbn from the picture. However,  there is a lot of work to be done to make this more robust/handle errors etc.